flask_summarization_app/requirements.txt
Flask
transformers
torch
torchaudio
pillow
